Alpine.data for timer state
